+++ src/g_shared/a_movingcamera.h
@@ -44,7 +44,9 @@ protected:
 
 	bool bPostBeginPlayCalled;
 	bool bActivateCalledBeforePostBeginPlay;
-	LONG lServerPrevNodeId, lServerCurrNodeId;
+
+	// [EP] TODO: remove the 'l' mark from the name of the variables which aren't LONG anymore
+	int lServerPrevNodeId, lServerCurrNodeId;
 	float fServerTime;
 public:
 	bool IsActive ();
